✨ Feature: Landingpage 분리. import 방식 변경, 호출 수정 (#37)
